refactor(cache-schema): rename collections impls to containers
