+++ crates/ssa-cache-schema/src/impls/containers.rs
