+++ crates/ssa-cache-schema/src/impls/mod.rs
@@ -1,4 +1,4 @@
-mod collections;
+mod containers;
 mod scalars;
 mod tuples;
